Consolidate duplicate back-end code
